Fix boolean and empty string value handling
Previously if a value was of type boolean, then the header entry would be added regardless of the boolean value (such that false would still get included).

Values of empty string would result in the header entry being output with no second argument. It makes more sense to skip the entry if the second half of the pair is empty.